apply security for connections at UI manipulation level

+++ core/java/iesi-rest-without-microservices/src/main/java/io/metadew/iesi/server/rest/connection/ConnectionsController.java
@@ -3,6 +3,8 @@ package io.metadew.iesi.server.rest.connection;
 import io.metadew.iesi.metadata.configuration.exception.MetadataAlreadyExistsException;
 import io.metadew.iesi.metadata.configuration.exception.MetadataDoesNotExistException;
 import io.metadew.iesi.metadata.definition.connection.key.ConnectionKey;
+import io.metadew.iesi.metadata.service.user.IESIPrivilege;
+import io.metadew.iesi.server.rest.configuration.security.IesiSecurityChecker;
 import io.metadew.iesi.server.rest.connection.dto.ConnectionDto;
 import io.metadew.iesi.server.rest.connection.dto.ConnectionDtoResourceAssembler;
 import io.metadew.iesi.server.rest.connection.dto.ConnectionDtoService;
@@ -17,6 +19,7 @@ import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.PagedModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
@@ -32,27 +35,29 @@ import java.util.List;
 @ConditionalOnWebApplication
 public class ConnectionsController {
 
-    private ConnectionService connectionService;
-    private ConnectionDtoService connectionDtoService;
-    private ConnectionDtoResourceAssembler connectionDtoResourceAssembler;
-    private PagedResourcesAssembler<ConnectionDto> connectionDtoPagedResourcesAssembler;
+    private final ConnectionService connectionService;
+    private final ConnectionDtoService connectionDtoService;
+    private final ConnectionDtoResourceAssembler connectionDtoResourceAssembler;
+    private final PagedResourcesAssembler<ConnectionDto> connectionDtoPagedResourcesAssembler;
+    private final IesiSecurityChecker iesiSecurityChecker;
 
     @Autowired
     ConnectionsController(ConnectionService connectionService,
                           ConnectionDtoService connectionDtoService,
                           ConnectionDtoResourceAssembler connectionDtoResourceAssembler,
-                          PagedResourcesAssembler<ConnectionDto> connectionDtoPagedResourcesAssembler) {
+                          PagedResourcesAssembler<ConnectionDto> connectionDtoPagedResourcesAssembler,
+                          IesiSecurityChecker iesiSecurityChecker) {
         this.connectionService = connectionService;
         this.connectionDtoService = connectionDtoService;
         this.connectionDtoResourceAssembler = connectionDtoResourceAssembler;
         this.connectionDtoPagedResourcesAssembler = connectionDtoPagedResourcesAssembler;
+        this.iesiSecurityChecker = iesiSecurityChecker;
     }
 
     @GetMapping("")
     @PreAuthorize("hasPrivilege('CONNECTIONS_READ')")
     public PagedModel<ConnectionDto> getAll(Pageable pageable, @RequestParam(required = false, name = "name") String name) {
         List<ConnectionFilter> connectionFilters = extractConnectionFilterOptions(name);
-        System.out.println("COUCOUCOUCOUCOUCOUCOUCOUC");
         Page<ConnectionDto> connectionDtoPage = connectionDtoService
                 .getAll(SecurityContextHolder.getContext().getAuthentication(),
                         pageable,
@@ -77,8 +82,8 @@ public class ConnectionsController {
     @PreAuthorize("hasPrivilege('CONNECTIONS_READ')")
     public ConnectionDto getByName(@PathVariable String name) {
         ConnectionDto connection = connectionDtoService.getByName(
-                SecurityContextHolder.getContext().getAuthentication(),
-                name)
+                        SecurityContextHolder.getContext().getAuthentication(),
+                        name)
                 .orElseThrow(() -> new MetadataDoesNotExistException(
                         new ConnectionKey(name, "")
                 ));
@@ -86,7 +91,7 @@ public class ConnectionsController {
     }
 
     @PostMapping("")
-    @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE')")
+    @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE', #connectionDto.securityGroupName)")
     public ResponseEntity<ConnectionDto> post(@Valid @RequestBody ConnectionDto connectionDto) {
         try {
             connectionService.createConnection(connectionDto);
@@ -99,7 +104,7 @@ public class ConnectionsController {
     }
 
     @PutMapping("")
-    @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE')")
+    @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE', #connectionDtos.![securityGroupName])")
     public HalMultipleEmbeddedResource<ConnectionDto> putAll(@Valid @RequestBody List<ConnectionDto> connectionDtos) throws MetadataDoesNotExistException {
         HalMultipleEmbeddedResource<ConnectionDto> halMultipleEmbeddedResource = new HalMultipleEmbeddedResource<>();
         connectionService.updateConnections(connectionDtos);
@@ -111,7 +116,7 @@ public class ConnectionsController {
     }
 
     @PutMapping("/{name}")
-    @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE')")
+    @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE', #connectionDto.securityGroupName)")
     public ConnectionDto put(@PathVariable String name, @RequestBody ConnectionDto connectionDto) throws MetadataDoesNotExistException {
         if (!connectionDto.getName().equals(name)) {
             throw new DataBadRequestException(name);
@@ -120,16 +125,18 @@ public class ConnectionsController {
         return connectionDtoResourceAssembler.toModel(connectionDto);
     }
 
-    @DeleteMapping("")
-    @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE')")
-    public ResponseEntity<?> deleteAll() {
-        connectionService.deleteAll();
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
     @DeleteMapping("/{name}")
     @PreAuthorize("hasPrivilege('CONNECTIONS_WRITE')")
     public ResponseEntity<?> deleteByName(@PathVariable String name) throws MetadataDoesNotExistException {
+        ConnectionDto connectionDto = connectionDtoService.getByName(null, name)
+                .orElseThrow(() -> new MetadataDoesNotExistException(new ConnectionKey(name, "")));
+
+        if (!iesiSecurityChecker.hasPrivilege(SecurityContextHolder.getContext().getAuthentication(),
+                IESIPrivilege.CONNECTIONS_MODIFY.getPrivilege(),
+                connectionDto.getSecurityGroupName())
+        ) {
+            throw new AccessDeniedException("User is not allowed to delete this connection");
+        }
         connectionService.deleteByName(name);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
